added more warnings to cmake, fixed some warnings

+++ src/components/denoising/gaussian-blur/GaussianBlurComponent.cpp
@@ -28,7 +28,7 @@ namespace components {
         void GaussianBlurComponent::applyDenoising() {
             const int half{ kernel_size / 2 };
 
-            std::vector<float> kernel(kernel_size * kernel_size);
+            std::vector<float> kernel(static_cast<size_t>(kernel_size * kernel_size));
             float sum{ 0.0f };
             for (int i{ -half }; i <= half; ++i) {
                 for (int j{ -half }; j <= half; ++j) {
@@ -36,7 +36,7 @@ namespace components {
                         std::exp(-(static_cast<float>(i * i + j * j)) /
                         (2.0f * sigma * sigma))
                     };
-                    kernel[(i + half) * kernel_size + (j + half)] = val;
+                    kernel[static_cast<size_t>((i + half) * kernel_size + (j + half))] = val;
                     sum += val;
                 }
             }
